volser: Remove ExtractVolId
volser was using its own function to extract a volume ID from a
filename string, and was using atol to do so. The ato* family of
functions can have problems with larger volume IDs, not to mention a
lack of error checking, so don't use it. Since we already have the
function VolumeNumber in the vol package to do the very same thing,
just use that instead.

Change-Id: I76ed65e4640cc42515091810689cd6e61f684ea0
Reviewed-on: http://gerrit.openafs.org/5594
Reviewed-by: Derrick Brashear <shadow@dementix.org>
Reviewed-by: Jeffrey Altman <jaltman@secure-endpoints.com>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
(cherry-picked from 87f969f82d12c8200ede3fd30a151e09ab0bf6ad)
Reviewed-on: http://gerrit.openafs.org/5917
Tested-by: Jeffrey Altman <jaltman@secure-endpoints.com>