Windows: Set new file index correctly during rename
Directory entries are required to have a file index which is used during
directory enumeration.  When inserting into a new directory we have to
update this from the target directory.

This code fixes a bug whereby it was being set from the source FCB, rather
than the target one.  On failure we now also reset the value to its old value.

Change-Id: I726dd2e7ecc84e4a8b28f1f551c89fc6e815d853
Reviewed-on: http://gerrit.openafs.org/5646
Tested-by: Jeffrey Altman <jaltman@secure-endpoints.com>
Reviewed-by: Jeffrey Altman <jaltman@secure-endpoints.com>