Windows: Look for 8.3 name when doing a rename
If we are doing a destructive rename we need to find whether the
target file exists.  This is done in the usual way (case sensitive
case insensitive), but the short name is not looked for.

This means that the rename of a file to a short name will not
supersede correctly; rather the service refuses the rename since
the target existed already.

This patch looks the target name up in the shortname tree if the
target name is short and all else has failed.

Change-Id: Ie054516a378edf52938d3465143a16107a99c0f6
Reviewed-on: http://gerrit.openafs.org/5652
Reviewed-by: Jeffrey Altman <jaltman@secure-endpoints.com>
Tested-by: Jeffrey Altman <jaltman@secure-endpoints.com>