Windows: Look for 8.3 name when doing a rename
If we are doing a destructive rename we need to find whether the
target file exists.  This is done in the usual way (case sensitive
case insensitive), but the short name is not looked for.

This means that the rename of a file to a short name will not
supersede correctly; rather the service refuses the rename since
the target existed already.

This patch looks the target name up in the shortname tree if the
target name is short and all else has failed.

Change-Id: Ib03a87e58a9f015000fc2bf3a4067b8683032a49
Reviewed-on: http://gerrit.openafs.org/5652
Reviewed-by: Jeffrey Altman <jaltman@secure-endpoints.com>
Tested-by: Jeffrey Altman <jaltman@secure-endpoints.com>
(cherry-picked from 241d3d54cebfc661cfaec6d21de091d906df3d08)
Reviewed-on: http://gerrit.openafs.org/5938