Windows: Directory Entry Processing
Memory corruption was detected due to improper handling
of directory entries and the short name b-tree during
directory revalidation and tear down.  The problem
resulted in a BSOD most frequently when roaming profiles
were written back to the file server or when multiple
clients were actively modifying the contents of a directory
simultaneously.

This patchset makes a number of changes:

 . rename processing no longer handles collisions locally.
   the file server replaces a target atomicly when there
   is a collision.

 . directory validation is no longer delayed until
   a handle is closed

 . data version numbers are checked for implied
   invalidation

Change-Id: Ib4b281bfb5da4b5a5ce735e85d85c50e15fbcde1
Reviewed-on: http://gerrit.openafs.org/5736
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Rod Widdowson <rdw@steadingsoftware.com>
Reviewed-by: Jeffrey Altman <jaltman@secure-endpoints.com>
Tested-by: Jeffrey Altman <jaltman@secure-endpoints.com>